requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.5
colorama==0.4.6
comtypes==1.2.0
distlib==0.3.7
filelock==3.12.2
Flask==2.3.2
frozenlist==1.3.3
gTTS==2.3.2
ibm-cloud-sdk-core==3.16.7
ibm-watson==7.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.26.5
platformdirs==3.10.0
PyAudio==0.2.13
pygame==2.5.0
PyJWT==2.7.0
pypiwin32==223
python-dateutil==2.8.2
python-dotenv==1.0.0
pyttsx3==2.90
requests==2.31.0
six==1.16.0
SpeechRecognition==3.10.0
tqdm==4.64.1
urllib3==1.26.14
virtualenv==20.24.2
websocket-client==1.1.0
Werkzeug==2.3.6
yarl==1.8.2
